Trait email_format::TryFrom

pub trait TryFrom<T>: Sized {
    type Err;
    fn try_from(T) -> Result<Self, Self::Err>;
}

Attempt to construct Self via a conversion (borrowed from rust std)

This TryFrom trait is defined in the rust std library but is behind a feature gate. We place it here so that people using stable compilers can still use our crate. In the future, the std trait should be used.

Associated Types

type Err

The type returned in the event of a conversion error.

Required Methods

fn try_from(T) -> Result<Self, Self::Err>

Performs the conversion.